requirements.txt
django==4.1.5
psycopg2==2.9.5
djangorestframework==3.14.0
markdown==3.3.7
django-filter==22.1